[skip ci] move the main three classes to distinguish from writeables
